Add generic type support.
- Refactors the annotation support for test_attrs
- Refactors errors and types into their own modules.
