+++ pyproject.toml
@@ -13,34 +13,23 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.5"
+python = "^3.4"
 python-dateutil = {version="^2.7", python = "<3.7"}
 
 [tool.poetry.dev-dependencies]
 pytest = "^4.1"
 attrs = "^18.2"
 pytest-cov = "^2.6"
-black = {version = "^19.3-beta.0",allows-prereleases = true,python = ">=3.7"}
-hypothesis = "^4.14"
+black = {version = "^19.3-beta.0", allow-prereleases = true, python = ">=3.7"}
+hypothesis = "^4"
+dataclasses = {version="0.4", python = ">=3.6,<3.7"}
+typing = {version = "^3.7", python = "<3.5"}
+typing-extensions = {version = "^3.7", python = "<3.8"}
 
 [tool.black]
 line-length = 88
-target-version = ["py35"]
+target-version = ["py34"]
 
-[tool.tox]
-legacy_tox_ini = """
-[tox]
-isolated_builds = true
-envlist = py35, py36, py37, py38
-skipsdist = true
-
-[testenv]
-deps =
-    poetry
-commands =
-    poetry install
-    poetry run pytest {posargs}
-"""
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
