+++ tests/common.py
@@ -1,3 +1,6 @@
+from importlib import import_module
+
+
 class Rules:
     def __init__(self, *rules):
         self.rules = rules
@@ -11,3 +14,24 @@ class Rules:
             return None
         else:
             raise RuntimeError("No rule for verb={}, typ={}".format(verb, typ))
+
+
+class SoftMod:
+    def __init__(self, *modules):
+        self.mods = mods = []
+        for name in modules:
+            try:
+                mods.append(import_module(name))
+            except ImportError:
+                pass
+
+    def __getattr__(self, name):
+        for mod in self.mods:
+            val = getattr(mod, name, None)
+            if val is not None:
+                return val
+        return None
+
+
+typing = SoftMod("typing", "typing_extensions")
+dataclasses = SoftMod("dataclasses")
